Update addLinks.py
This program now supports shapefiles which are in a zipfile, but the shapefile in the zipfile must be the one that was created by my other tools in order to work properly. 

Currently, there is no error handling if the zip file doesn't work.  Also, the latest version of this tool isn't implemented by the ArcGIS toolbox which uses these scripts. I hope to change that soon.