Major Changes, ahoy
These modifications allow individuals who re-use these tools to use the
spatial index zip archive, an output for all three of the major tools in
this package, as an input rather than unpackaging the shapefile it
containes. Documentation was also modified for the tool which adds
website information after a successful instance of the "BrechinBulk
georeference and index" tool.